+++ src/applications/trading-app/trade-store.js
@@ -144,28 +144,31 @@ export default class TradeStore {
 
   addItem(newItem, { quantity = false, currency = false } = {}) {
 
-    const items = !currency ? get(this.leftTraderItems) : get(this.leftTraderItemCurrencies);
+    const items = !currency
+      ? get(this.leftTraderItems)
+      : get(this.leftTraderItemCurrencies);
 
     const item = Utilities.findSimilarItem(items, newItem)
 
     const maxQuantity = game.user.isGM ? Infinity : Utilities.getItemQuantity(newItem);
 
-    if (!item) {
+    if (item && Utilities.canItemStack(item)) {
+      if (item.quantity >= maxQuantity) return;
+      item.quantity = Math.min(quantity ? quantity : item.quantity + 1, maxQuantity);
+      item.newQuantity = item.quantity;
+      item.maxQuantity = maxQuantity;
+    } else if (!item) {
       items.push({
         id: newItem.id,
         name: newItem.name,
         img: newItem?.img ?? "",
+        type: newItem?.type,
         currency: currency,
         quantity: quantity ? quantity : 1,
         newQuantity: quantity ? quantity : 1,
         maxQuantity: maxQuantity,
         data: newItem instanceof Item ? newItem.toObject() : newItem
       })
-    } else {
-      if (item.quantity >= maxQuantity) return;
-      item.quantity = Math.min(quantity ? quantity : item.quantity + 1, maxQuantity);
-      item.newQuantity = item.quantity;
-      item.maxQuantity = maxQuantity;
     }
 
     if (!currency) {